Adjusted grades: easier master rank, harder As and Bs
Adjusted all grade criteria with an emphasis on making it easier to reach the
highest grades, and harder to reach the lowest grades:

M:   100.0% -> 95.0% (-5%)
SSS:  85.4% -> 81.1% (-5%)
SS:   69.2% -> 65.7% (-5%)
S-:   21.7% -> 20.6% (-5%)
A-:    3.6% ->  8.8% (+245%)
B-:    1.4% ->  4.7% (+347%)

I originally though M-ranks should be for TAS-level play, but it's unfun to have
the grading system be that strict. I think it's better to give players an M-rank
for expert-level play, and then people who want to strive for TAS-level play can
push their scores past 'M' if they want.

I still think it's nice to grade novices gently and give them a 'B'
grade for a bad score, but based on player feedback I was overdoing it. The
criteria for a 'B' or 'A' grade are still low, but they're higher than before.