Bug: Progress board was not centered.
The progress board animations hard-coded x-coordinates like "422" but
these x-coordinates were only valid for a specific resolution. As a
result, the animations would animate the progress board into an
uncentered position for most resolutions.

I've introduced a new 'swoosher' node in the ProgressBoard which has an
X-coordinate of 0,0, and will appear centered regardless of the
resolution.

Closes #2937.