Bug: Creature editor color buttons didn't return focus
Pressing a Creature Editor color button and then closing the popup
lost focus. The focus was not returned to the color button afterward.

This bug was confusing to track down, but it only happens if the
ColorPickerPopup has at least one preset, and if the resulting preset
button grabs focus.

Updated CreatureColorButtonDemo to include multiple palettes.