Preempted all get_tree calls with is_inside_tree() checks
A user experienced a game crash after the end of day in career mode with the following message logged:

ERROR: Condition "!data.tree" is true. Returned: nullptr
   at: Node::get_tree (.\scene/main/node.h:333) - Condition "!data.tree" is true. Returned: nullptr

It's possible get_tree() could cause this error in some edge cases --
for example, if someone called 'LandmarkLines.refresh()' after it was
removed from the scene tree. I've preempted all get_tree() calls with
checks to ensure the node is still in the tree.

Note: I did not add checks to most calls in `_ready` or `_enter_tree`
since those callbacks are only called when the node is in the tree.