Decreased 'avoid repeated sfx' threshold from 200 ms to 20 ms
The high threshold was making the hover sound effects a little strange
and sparse during the menus. Circling the mouse quickly would not play
very many hover sounds.

This threshold was accidentally increased in a099b976 when fixing overlapping
audio for the training menu unlock cheat. This change affected the entire game,
not just the cheat sound effect, and increased the threshold very high.