Added 'ViewportPurger' script to avoid crashes.
Godot crashes sometimes when freeing a node with a ViewportTexture property. To
prevent this, we manually set any ViewportTexture properties to null before the
node is freed. ViewportPurger accepts a list of viewport path properties
relative to a particular root node. The string syntax is a NodePath followed by
one or more colon-delimited properties.