Bug: Autopilot only worked at 1024x600
Autopilot only worked if the window was exactly 1024x600 pixels (the
default size.) If the window was maximized or resized, the mouse events
would click the wrong parts of the screen.

I've changed Autopilot's press_mouse and release_mouse logic to apply
the viewport transform so that it works in windows of any size.