Easier difficulties have easier success criteria.
Before, instead of easier boss levels requiring something like "Get $300 in 3
minutes", they were adjusted to something like "Get $300 in 1 minute" which is
much, much harder. Shortening normal levels made them easier, but
shortening boss level made it much harder to meet the requirements.

Now, we scale the success criteria accordingly to "Get $100 in 1
minute", but also make an additional adjustment because novices won't be
able to score as many points per minute as experts. So the final result
is something silly like "Get $10 in 1 minute". If players think it is
too easy, there are harder difficulties. The goal of Zen mode is to open
up the game to all players.